repo::open: remove dead code, free buffers
